requirements.txt
llama-index
pypdf
pandas
python-dotenv
llama-index-experimental
llama-index-llms-huggingface
llama-index-llms-groq